# api/gen/go/server/service/v1/i_sandbox_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: server/service/v1/i_sandbox.proto

package serverpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-admin/api/gen/go/sandbox/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSandboxServiceCreate = "/server.service.v1.SandboxService/Create"
const OperationSandboxServiceDelete = "/server.service.v1.SandboxService/Delete"
const OperationSandboxServiceGet = "/server.service.v1.SandboxService/Get"
const OperationSandboxServiceGetPassword = "/server.service.v1.SandboxService/GetPassword"
const OperationSandboxServiceList = "/server.service.v1.SandboxService/List"
const OperationSandboxServiceRestart = "/server.service.v1.SandboxService/Restart"
const OperationSandboxServiceStart = "/server.service.v1.SandboxService/Start"
const OperationSandboxServiceStop = "/server.service.v1.SandboxService/Stop"
const OperationSandboxServiceUpdate = "/server.service.v1.SandboxService/Update"

type SandboxServiceHTTPServer interface {
	// Create 创建沙箱（需先调用资源服务分配GPU）
	Create(context.Context, *v11.CreateSandboxRequest) (*emptypb.Empty, error)
	// Delete 删除沙箱（销毁，需先调用资源服务释放GPU）
	Delete(context.Context, *v11.DeleteSandboxRequest) (*emptypb.Empty, error)
	// Get 沙箱详情
	Get(context.Context, *v11.GetSandboxRequest) (*v11.Sandbox, error)
	// GetPassword 获取沙箱访问密码
	GetPassword(context.Context, *v11.GetSandboxPasswordRequest) (*v11.GetSandboxPasswordResponse, error)
	// List 沙箱列表
	List(context.Context, *v1.PagingRequest) (*v11.ListSandboxResponse, error)
	// Restart 重启沙箱
	Restart(context.Context, *v11.RestartSandboxRequest) (*emptypb.Empty, error)
	// Start 启动沙箱
	Start(context.Context, *v11.StartSandboxRequest) (*emptypb.Empty, error)
	// Stop 停止沙箱
	Stop(context.Context, *v11.StopSandboxRequest) (*emptypb.Empty, error)
	// Update 更新沙箱
	Update(context.Context, *v11.UpdateSandboxRequest) (*emptypb.Empty, error)
}

func RegisterSandboxServiceHTTPServer(s *http.Server, srv SandboxServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/server/v1/sandboxes", _SandboxService_List2_HTTP_Handler(srv))
	r.GET("/server/v1/sandboxes/{id}", _SandboxService_Get2_HTTP_Handler(srv))
	r.POST("/server/v1/sandboxes", _SandboxService_Create2_HTTP_Handler(srv))
	r.PUT("/server/v1/sandboxes/{id}", _SandboxService_Update2_HTTP_Handler(srv))
	r.DELETE("/server/v1/sandboxes/{id}", _SandboxService_Delete2_HTTP_Handler(srv))
	r.POST("/server/v1/sandboxes/{id}:restart", _SandboxService_Restart1_HTTP_Handler(srv))
	r.POST("/server/v1/sandboxes/{id}:start", _SandboxService_Start1_HTTP_Handler(srv))
	r.POST("/server/v1/sandboxes/{id}:stop", _SandboxService_Stop1_HTTP_Handler(srv))
	r.POST("/server/v1/sandboxes/{id}/password", _SandboxService_GetPassword0_HTTP_Handler(srv))
}

func _SandboxService_List2_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListSandboxResponse)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Get2_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetSandboxRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Sandbox)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Create2_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateSandboxRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Update2_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateSandboxRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Delete2_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteSandboxRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Restart1_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.RestartSandboxRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceRestart)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Restart(ctx, req.(*v11.RestartSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Start1_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.StartSandboxRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceStart)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Start(ctx, req.(*v11.StartSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_Stop1_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.StopSandboxRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceStop)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Stop(ctx, req.(*v11.StopSandboxRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SandboxService_GetPassword0_HTTP_Handler(srv SandboxServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetSandboxPasswordRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSandboxServiceGetPassword)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPassword(ctx, req.(*v11.GetSandboxPasswordRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.GetSandboxPasswordResponse)
		return ctx.Result(200, reply)
	}
}

type SandboxServiceHTTPClient interface {
	// Create 创建沙箱（需先调用资源服务分配GPU）
	Create(ctx context.Context, req *v11.CreateSandboxRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Delete 删除沙箱（销毁，需先调用资源服务释放GPU）
	Delete(ctx context.Context, req *v11.DeleteSandboxRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 沙箱详情
	Get(ctx context.Context, req *v11.GetSandboxRequest, opts ...http.CallOption) (rsp *v11.Sandbox, err error)
	// GetPassword 获取沙箱访问密码
	GetPassword(ctx context.Context, req *v11.GetSandboxPasswordRequest, opts ...http.CallOption) (rsp *v11.GetSandboxPasswordResponse, err error)
	// List 沙箱列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListSandboxResponse, err error)
	// Restart 重启沙箱
	Restart(ctx context.Context, req *v11.RestartSandboxRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Start 启动沙箱
	Start(ctx context.Context, req *v11.StartSandboxRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Stop 停止沙箱
	Stop(ctx context.Context, req *v11.StopSandboxRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Update 更新沙箱
	Update(ctx context.Context, req *v11.UpdateSandboxRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type SandboxServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewSandboxServiceHTTPClient(client *http.Client) SandboxServiceHTTPClient {
	return &SandboxServiceHTTPClientImpl{client}
}

// Create 创建沙箱（需先调用资源服务分配GPU）
func (c *SandboxServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateSandboxRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/sandboxes"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSandboxServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除沙箱（销毁，需先调用资源服务释放GPU）
func (c *SandboxServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteSandboxRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/sandboxes/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSandboxServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 沙箱详情
func (c *SandboxServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetSandboxRequest, opts ...http.CallOption) (*v11.Sandbox, error) {
	var out v11.Sandbox
	pattern := "/server/v1/sandboxes/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSandboxServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetPassword 获取沙箱访问密码
func (c *SandboxServiceHTTPClientImpl) GetPassword(ctx context.Context, in *v11.GetSandboxPasswordRequest, opts ...http.CallOption) (*v11.GetSandboxPasswordResponse, error) {
	var out v11.GetSandboxPasswordResponse
	pattern := "/server/v1/sandboxes/{id}/password"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSandboxServiceGetPassword))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 沙箱列表
func (c *SandboxServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListSandboxResponse, error) {
	var out v11.ListSandboxResponse
	pattern := "/server/v1/sandboxes"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSandboxServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Restart 重启沙箱
func (c *SandboxServiceHTTPClientImpl) Restart(ctx context.Context, in *v11.RestartSandboxRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/sandboxes/{id}:restart"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSandboxServiceRestart))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Start 启动沙箱
func (c *SandboxServiceHTTPClientImpl) Start(ctx context.Context, in *v11.StartSandboxRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/sandboxes/{id}:start"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSandboxServiceStart))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Stop 停止沙箱
func (c *SandboxServiceHTTPClientImpl) Stop(ctx context.Context, in *v11.StopSandboxRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/sandboxes/{id}:stop"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSandboxServiceStop))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新沙箱
func (c *SandboxServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateSandboxRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/sandboxes/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSandboxServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}