# app/server/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: AIMS GO API
    description: AIMS GO API
    contact:
        name: wang.qiyuan
        url: https://github.com/aims-go/aims-go
        email: wang.qiyuan@21vianet.com
    license:
        name: MIT License
        url: https://github.com/aims-go/aims-go/blob/master/LICENSE
    version: "1.0"
paths:
...
    /server/v1/sandboxes:
        get:
            tags:
                - SandboxService
            description: 沙箱列表
            operationId: SandboxService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListSandboxResponse'
        post:
            tags:
                - SandboxService
            description: 创建沙箱（需先调用资源服务分配GPU）
            operationId: SandboxService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateSandboxRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /server/v1/sandboxes/{id}:
        get:
            tags:
                - SandboxService
            description: 沙箱详情
            operationId: SandboxService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Sandbox'
        put:
            tags:
                - SandboxService
            description: 更新沙箱
            operationId: SandboxService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateSandboxRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - SandboxService
            description: 删除沙箱（销毁，需先调用资源服务释放GPU）
            operationId: SandboxService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: removeHistory
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content: {}
    /server/v1/sandboxes/{id}/password:
        post:
            tags:
                - SandboxService
            description: 获取沙箱访问密码
            operationId: SandboxService_GetPassword
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetSandboxPasswordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetSandboxPasswordResponse'
    /server/v1/sandboxes/{id}:restart:
        post:
            tags:
                - SandboxService
            description: 重启沙箱
            operationId: SandboxService_Restart
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RestartSandboxRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /server/v1/sandboxes/{id}:start:
        post:
            tags:
                - SandboxService
            description: 启动沙箱
            operationId: SandboxService_Start
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StartSandboxRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /server/v1/sandboxes/{id}:stop:
        post:
            tags:
                - SandboxService
            description: 停止沙箱
            operationId: SandboxService_Stop
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StopSandboxRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreateSandboxRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Sandbox'
                password:
                    writeOnly: true
                    type: string
                    description: 访问密码（仅输入，不返回）
                flavor_name:
                    type: string
                    description: 指定运行规格
            description: 创建沙箱请求
...
        GetSandboxPasswordRequest:
            type: object
            properties:
                id:
                    type: string
            description: 沙箱详情请求
        GetSandboxPasswordResponse:
            type: object
            properties:
                password:
                    type: string
                    description: 访问密码
            description: 沙箱详情请求
...
        ListSandboxResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Sandbox'
                total:
                    type: string
            description: 沙箱列表响应
...
        Quota:
            type: object
            properties:
                cpu:
                    type: string
                memory:
                    type: string
                storage:
                    type: string
                gpu:
                    type: string
...
        RestartSandboxRequest:
            type: object
            properties:
                id:
                    type: string
            description: 重启沙箱请求
...
        Sandbox:
            required:
                - name
            type: object
            properties:
                id:
                    type: string
                    description: 沙箱ID
                name:
                    maxLength: 20
                    pattern: ^[a-z0-9]{1,20}$
                    type: string
                    description: 沙箱名称，仅允许小写字母和数字，长度不超过20个字符
                modelId:
                    type: string
                    description: 关联模型ID
                modelName:
                    type: string
                    description: 模型名称
                tenantName:
                    type: string
                    description: 所属租户
                status:
                    enum:
                        - UNSPECIFIED
                        - CREATING
                        - RUNNING
                        - STOPPED
                        - RESTARTING
                        - ERROR
                        - DELETING
                        - EXPIRED
                    type: string
                    description: 状态
                    format: enum
                config:
                    example: null
                    type: object
                    additionalProperties:
                        type: string
                description:
                    type: string
                    description: 描述
                remark:
                    type: string
                    description: 备注
                quota:
                    $ref: '#/components/schemas/Quota'
                deploymentName:
                    readOnly: true
                    type: string
                    description: 关联的K8s Deployment名称
                url:
                    type: string
                    description: url
                commandMapping:
                    type: object
                    additionalProperties:
                        type: string
                    description: '命令映射，用于不同环境下的命令适配，例如 {''start'': ''python app.py'', ''jupyter'': ''jupyter notebook --ip=}'
                env:
                    type: object
                    additionalProperties:
                        type: string
                    description: '环境变量，键值对形式，例如 {''ENV_VAR1'': ''value1'', ''ENV_VAR2'': ''value2''}'
                startedAt:
                    type: string
                    description: 启动时间
                    format: date-time
                finishedAt:
                    type: string
                    description: 结束时间
                    format: date-time
                expirationMinutes:
                    type: string
                    description: 过期时间
                billingDurationSec:
                    type: string
                    description: 计费时长
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 沙箱详情信息
        StartSandboxRequest:
            type: object
            properties:
                id:
                    type: string
            description: 启动沙箱请求
        StopSandboxRequest:
            type: object
            properties:
                id:
                    type: string
            description: 停止沙箱请求
...
        UpdateSandboxRequest:
            type: object
            properties:
                id:
                    type: string
                data:
                    $ref: '#/components/schemas/Sandbox'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新沙箱请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...